+++ GRDB/Core/Support/Foundation/Data.swift
@@ -8,17 +8,16 @@ import Foundation
 /// Data is convertible to and from DatabaseValue.
 extension Data : DatabaseValueConvertible, StatementColumnConvertible {
     public init(sqliteStatement: SQLiteStatement, index: Int32) {
-        let bytes = unsafeBitCast(sqlite3_column_blob(sqliteStatement, Int32(index)), to: UnsafePointer<UInt8>.self)
-        let count = Int(sqlite3_column_bytes(sqliteStatement, Int32(index)))
-        self.init(bytes: bytes, count: count) // copy bytes
+        if let bytes = sqlite3_column_blob(sqliteStatement, Int32(index)) {
+            let count = Int(sqlite3_column_bytes(sqliteStatement, Int32(index)))
+            self.init(bytes: bytes, count: count) // copy bytes
+        } else {
+            self.init()
+        }
     }
     
     /// Returns a value that can be stored in the database.
     public var databaseValue: DatabaseValue {
-        // SQLite cant' store zero-length blobs.
-        guard count > 0 else {
-            return .null
-        }
         return DatabaseValue(storage: .blob(self))
     }